reinstate mathspeak for english

+++ ts/speech_rules/clearspeak_preference_string.ts
@@ -58,7 +58,7 @@ export const PREFERENCES = new DynamicProperties({
     // Reciprocal is French
     'Reciprocal'
   ],
-  Inference: ['Auto', 'Short'],
+  Inference: ['Auto', 'Long'],
   ImpliedTimes: ['Auto', 'MoreImpliedTimes', 'None'],
   Log: ['Auto', 'LnAsNaturalLog'],
   Matrix: [
